services/ml_service/requirements.txt
fastapi
uvicorn
pandas
pickle4
scikit-learn
prometheus_fastapi_instrumentator
psycopg2-binary
python-multipart